requirements.txt
docopt==0.6.2
stomp.py==8.0.1
